added apache log file parser from mvar
