+++ composer.json
@@ -18,6 +18,7 @@
     "symfony/dependency-injection": "~2.6",
     "symfony/config": "~2.6",
     "symfony/yaml": "~2.6",
-    "symfony/event-dispatcher": "~2.6"
+    "symfony/event-dispatcher": "~2.6",
+    "mvar/apache2-log-parser": "~1.0"
   }
 }
